backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-dotenv==1.0.0
stripe==7.6.0
requests==2.31.0
psycopg2-binary==2.9.9 